requirements.txt
# 基础依赖
# 需要系统安装ffmpeg二进制工具

# 视频帧去重功能依赖
requests>=2.25.0          # 用于调用Jina API
numpy>=1.21.0             # 用于向量计算
Pillow>=8.0.0             # 用于图片处理
scipy>=1.7.0              # 用于高级数学计算（可选）

# 图文混排笔记功能依赖
reportlab>=3.6.0          # 用于PDF生成

# 可选：用于更高级的视频处理
# opencv-python>=4.5.0
# moviepy>=1.0.0
